pull socket.io logic out from Editor component into a helper and stores hashes in treeNode model
