+++ src/server/models/user.js
@@ -1,5 +1,3 @@
-// src/models/user.js
-
 const mongoose = require('mongoose');
 const bcrypt = require('bcrypt-nodejs');
